move delete logic to another component
